+++ packages/shared-base/src/legacy-settings/settings.ts
@@ -1,7 +1,10 @@
 import { isEqual } from 'lodash-es'
 import { Appearance, LanguageOptions } from '@masknet/public-api'
 import { createGlobalSettings, createBulkSettings } from './createSettings.js'
-import { BooleanPreference, EnhanceableSite, ExtensionSite, NetworkPluginID, updateLanguage } from '../index.js'
+import { updateLanguage } from '../i18n/index.js'
+import { NetworkPluginID } from '../Plugin/index.js'
+import { EnhanceableSite, ExtensionSite } from '../Site/types.js'
+import { BooleanPreference } from '../legacy-settings/types.js'
 
 export const languageSettings = createGlobalSettings<LanguageOptions>('language', LanguageOptions.__auto__)
 languageSettings.addListener(updateLanguage)
